Cleanup relocation.
