+++ service/internal/graph/graph_test.go
@@ -2614,4 +2614,3 @@ func testGraphBuildErrors(t *testing.T) {
 		})
 	}
 }
-
